feat: se agrego el modelo de random forest para realizar predicciones
Se entreno e implemento el modelo de random forest con un r2 de 0.7

+++ ML/schema.py
@@ -13,5 +13,5 @@ class Regiones(str,Enum):
 class Datainput(BaseModel):
     habs: int
     baños: int
-    m2: int
+    m2: float
     region: Regiones
